nbody: fix case where force=0 leads to NaN's and subsequent quadtree crash
In rare cases the total force on a node can be 0.  This leads to infinite
dt, which leads to NaN's for the node's x and y position.  Due to centring
of all nodes, the NaN contaminates all nodes.  This leads to a quadtree
with NaN bounding box and NaN for all nodes, and all nodes try to go in
exactly the same location in the quadtree, leading to infinite depth of the
quadtree.  This crashes with a segfault due to a massive stack depth.