tiles: allow to associate a region label with a paper id
The graph gradually shifts and changes shape over time, so region labels
(which were hand placed at the beginning) are gradually out of place.  This
patch makes it so you can optionally associate a region label with a paper
id, and if you do that then the region label position tracks the paper
position.